mfields/cuda: mv StorageUniquePtr -> kg/Storage.h
